[Improvement]: support access Tables tab only for display table info only
